control/requirements.txt
django == 4.0.3
django-environ == 0.8.1
django-tailwind == 3.1.1
pyserial == 3.5